Fix finefile.toml always being included
If we specified a different config file using
-f, finefile.toml was still included as the first
config file. The intention was that finefile.toml
would only be automatically included if no config
files were specified.

Fix by removing the default value for the
parameter. Instead we check for config files at
the point of use, and add finefile.toml if the
seq is empty.