Don't run hyperfine when there are no steps to run
Currently there can be no setup/cleanup etc or
command to run, and we just run hyperfine with
a command of "true". But this is wasteful and a
bit confusing.

Address by detecting when we have
nothing to run and skipping hyperfine entirely.

Since we don't run hyperfine, we also won't have
a JSON export so we skip reading that.